Added basic skeleton of the server. twisted is mind-bendingly cool.
At the moment, the server has one command, test, and expects the client to
have one command, foo (don't ask).

Starting the server:

[tom@eee server]$ python run.py
Running on port 11235...

Testing with telnet:

[tom@eee ~]$ telnet localhost 11235
Trying 127.0.0.1...
Connected to localhost.
Escape character is '^]'.
>>> [12345, "test"]
[1, "foo", "hello"]
>>> [1, "return", 0, "hi"]
[12345, "return", 0, "you said: hi"]

It works, but it's middle-of-the-night code, and needs plenty of polishing,
and the actual usefull bits need adding. Feel free to hack away and make
things better in a general way.