Fix memory leak.
I'm pretty sure that this was caused by a bug in loopbackAsync. This is
'fixed' by connecting the meta client via a random UNIX socket. It
should be possible to fix this by basically re-implementing
loopbackAsync in a not-terrible way, but that might be more work than it
looks.

tempfile.mktemp is used even though it's deprecated, because it's still
the easiest way to do this, which is stupid. Also twisted uses it, so
it's justified, right.